+++ ACViewer/Physics/Common/BuildingObj.cs
@@ -30,7 +30,7 @@ namespace ACE.Server.Physics.Common
             CurCell = newCell;
         }
 
-        public void add_to_stablist(ref List<ushort> blockStabList, ref uint maxSize, ref uint stabNum)
+        public void add_to_stablist(ref List<uint> blockStabList, ref uint maxSize, ref uint stabNum)
         {
             foreach (var portal in Portals)
                 portal.add_to_stablist(ref blockStabList, ref maxSize, ref stabNum);
